airtime-liquidsoap should also start Monit
The /etc/init.d/airtime-liquidsoap script is called by Debian/Ubuntu with the 'start' (or 'stop') argument when the machine boot (or shutdown).  The current 'start' argument does not launch monit. Therefore, by default, monit is not required to monitor the started process. Shouldn't it be the case by default?

This commit makes 'start' to be an alias to 'start-with-monit', and add a new 'start-whitout-monit' option.